+++ src/index.ts
@@ -2,3 +2,5 @@ export * from './exec';
 export * from './execFile';
 export * from './fork';
 export * from './spawn';
+export * from './operators';
+export * from './utils';