+++ src/util.ts
@@ -1,4 +1,5 @@
 import {ChildProcess} from 'child_process';
+import {Observable, Subject} from 'rxjs';
 
 export function killProc(proc: ChildProcess) {
   if (proc.stdout) {
@@ -11,4 +12,12 @@ export function killProc(proc: ChildProcess) {
 
   proc.removeAllListeners();
   proc.kill('SIGKILL');
+}
+
+export function spawnEnd(spawnObservable: Observable<any>) {
+  const sbj = new Subject<void>();
+
+  spawnObservable.subscribe(undefined, err => sbj.error(err), () => sbj.next());
+
+  return sbj;
 }