+++ test/fixtures/stdoutMultiple.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+echo HELLO
+
+sleep 0.05
+
+echo WORLD