+++ CHANGELOG.md
@@ -1,1 +1,12 @@
+# [0.0.3] (2018-10-13)
+
+- change package name to `rxjs-shell`
+
+### Features
+
+- add `util/spawnEnd`: to know when `spawn` stream completed.
+- add `operators/trim`: trim output buffer or string contents.
+- add `operators/printBuf`: syntax sugar of `tap(buf => process.stdout.write(buf))`.
+
 # [0.0.2](https://github.com/johnny-mh/rxjs-shell-operators/commit/d249d3570dcc6d87d200aae4570c621a90aafdeb) (2018-10-10)
+